GPT4-Version/requirements.txt
Flask==3.0.3
gunicorn==20.1.0
diffusers==0.27.2
moviepy==1.0.3
gTTS==2.5.1
supabase==2.4.6
requests==2.31.0
Pillow==10.3.0
